refactor: move class function inside Trainer
