+++ mindnlp/engine/utils.py
@@ -505,28 +505,6 @@ def speed_metrics(split, start_time, num_samples=None, num_steps=None, num_token
         result[f"{split}_tokens_per_second"] = round(tokens_per_second, 3)
     return result
 
-def _get_learning_rate(self):
-    r"""
-    This function retrieves the learning rate used by the optimizer.
-    
-    Args:
-        self: An instance of the class containing the optimizer and learning rate scheduler.
-    
-    Returns:
-        The learning rate value (float) used by the optimizer.
-    
-    Raises:
-        None.
-    """
-    if isinstance(self.lr_scheduler, optim.lr_scheduler.ReduceLROnPlateau):
-        last_lr = self.optimizer.param_groups[0]["lr"]
-    else:
-        last_lr = self.lr_scheduler.get_last_lr()[0]
-    if ops.is_tensor(last_lr):
-        last_lr = last_lr.item()
-    return last_lr
-
-
 def find_batch_size(tensors):
     """
     Find the first dimension of a tensor in a nested list/tuple/dict of tensors.
